requirements.txt
pyproj==1.9.3
python-geoip-geolite2==2014.0207
lxml==3.3.2
pojson==0.5
